refactor: restructure Go test files by target class name
