+++ api/internal/api/media.go
@@ -10,8 +10,9 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/labstack/echo/v4"
 	"point-api/internal/services"
+
+	"github.com/labstack/echo/v4"
 )
 
 type MediaHandler struct {
