+++ api/internal/api/middleware_test.go
@@ -1,0 +1,12 @@
+package api
+
+import "testing"
+
+func TestMiddleware_ExtractIDNil(t *testing.T) {
+	if id := extractUserID(nil); id != 0 {
+		t.Errorf("expected 0, got %d", id)
+	}
+	if id := extractSessionID(nil); id != 0 {
+		t.Errorf("expected 0, got %d", id)
+	}
+}
