+++ api/internal/api/nav_menu_test.go
@@ -1,0 +1,12 @@
+package api
+
+import (
+	"point-api/internal/services"
+	"testing"
+)
+
+func newNavMenuHandler(t *testing.T) *NavMenuHandler {
+	t.Helper()
+	repo := setupTestDB(t)
+	return NewNavMenuHandler(services.NewSettingsService(repo))
+}
