+++ api/internal/api/tags_threshold_test.go
@@ -8,9 +8,10 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/labstack/echo/v4"
 	"point-api/internal/models"
 	"point-api/internal/services"
+
+	"github.com/labstack/echo/v4"
 )
 
 func TestTagHandler_MinPostsThreshold(t *testing.T) {
